terraform/cloud_functions/oauth_login/requirements.txt
google-cloud-bigquery==3.11.4
PyJWT==2.8.0
flask==3.0.0
google-auth==2.23.4
requests==2.31.0
